feat: add an image module photos of the form
